fix overlaykey constructor & robust xml parsing
original Overlay key constructor shift the bit buffer to the high/left
end. but due to calculation related issue, it will malfunction when the
bits length are small. so just delete the offset it used to shift this
process. xml parsing some time will return an empty object which, if you
continue access its child/sibling, will prompt an BAD ACCESS error to
the OS, so just use if(child) to guarantee it is not empty, if it is
then return a Unspec-key as overlay key.